separate signals handling, name thread
I pulled the signals handling code out of engine; it has very few
dependencies, and I noticed that signal_hook is compiled in both
windows and unix builds. To help make it easier to statically
verify compilation on windows, I removed the conditional
compilation blocks and pulled the decision into a compile-time
decided const if {} block. This keeps the code there to satisfy
clippy w.r.t. symbol use.

This does have the side effect of also creating an active_config_path
pointer whether or not it will be used. It's benign and would be
dropped immediately on a windows non-websocket build. It also
names the signal thread. I kept the thread name below 16 characters
so it renders nicely on htop et al.
